services/signal/pyproject.toml
[project]
name = "nexusquant-signal"
version = "0.0.0"
description = "Signal service: rolling indicators for watchlist symbols."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "alpaca-py>=0.30",
    "pytz>=2024.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexusquant_signal"]
